third_party/blink/web_tests/platform/generic/fast/events/onunload-not-on-body-expected.txt

CONSOLE ERROR: Blocked alert('unload') during unload.
You should only see one unload alert appear.
